requirements.txt
pyqrcode
pillow
reportlab
webcolors
